Introduction
Tennis is a popular sport with a long history of competitive tournaments and rankings. A player's ranking is an essential metric used to evaluate their performance and overall skill level. While several factors can impact a player's ranking, such as age, seasons played, current rank, and prize money, it can be challenging to determine the best approach for predicting a player's best rank accurately.

In this proposal, we aim to predict a tennis player's best rank based on their age, seasons played, current rank, and prize money. We plan to use K-nearest neighbor regression, a machine learning algorithm, to analyze the data and make predictions. The goal of this project is to determine whether these variables have a direct or indirect impact on a player's best rank and which ones are most influential in making accurate predictions.

To accomplish this, we will use a dataset containing information on various tennis players, including their age, best rank, prize money, current rank, and seasons played. We will use five-fold cross-validation to estimate the smallest error in our model and then evaluate the optimal parameter value using the standard testing dataset. By analyzing the data and using visualization techniques such as scatter plots, we hope to gain valuable insights into the factors that influence a player's ranking in tennis.